pmt-ptw

+++ eclipse-collections/src/main/java/org/eclipse/collections/impl/bag/immutable/ImmutableArrayBag.java
@@ -84,7 +84,7 @@ public class ImmutableArrayBag<T>
             public void value(T each, int count)
             {
                 newKeys[this.index] = each;
-                newCounts[this.index] = count;
+                newCounts[this.index] = ((Integer)(count)).intValue();
                 this.index++;
             }
         });
